services/health_mongo/requirements.txt
google-cloud-error-reporting==1.0.*
google-cloud-bigquery==2.4.*
firebase-admin==5.0.*
pymongo==3.11.*